% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faceplot.R
\name{colorbar}
\alias{colorbar}
\title{colorpalette generation}
\usage{
colorbar(vec, col.pal = colorRampPalette(c("blue", "red")),
  output = NULL)
}
\arguments{
\item{vec}{A vector. Each of its elements will be assigned a RGB value from the color palette}

\item{col.pal}{A list of colors or a colorRampPalette object.
The default is \code{colorRampPalette(c('blue','red'))}}

\item{output}{the output path. the default is NULL.
if output is NULL, the color bar will not be printed to a file.}
}
\value{
a vector of colors
}
\description{
The function is used to genrate color palette, which will used for \code{visualed}
to render mesh object.
}
\examples{
colorbar(1:10,c('#377EB8', '#4DAF4A','white', '#984EA3', '#FF7F00'))

}
