% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\docType{data}
\name{geno}
\alias{geno}
\title{Genotype Data}
\format{An object of list with a Fromal class \code{'SnpMatrix'} and a data.frame}
\usage{
data(geno)
}
\description{
The data is a list with a SnpMatrix `genotypes` (2000 rows, 50 columns) and a data frame `map`.
It should be used in the \code{magpa} function to test multivariate correlation.
}
\details{
\itemize{
  \item genotypes
  \item map
}
}
\examples{
data(geno)
snps <- geno$genotypes
}
\keyword{datasets}
