% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpa.R
\name{magpa}
\alias{magpa}
\title{Multivariate Analysis of Genotype–Phenotype Association}
\usage{
magpa(geno, pheno, pca = FALSE, rm.na = TRUE)
}
\arguments{
\item{geno}{string; a file prefix name of plink output (.bed, .bim, .fam),
list; an object read by \code{read.plink} of the **snpStats** package,
or a SnpMatrix}

\item{pheno}{a matrix with n rows of sample and m columns of features}

\item{pca}{boolean. perform principal component analysis on the phenotype?
the default is FALSE}

\item{rm.na}{boolean. remove missing values during analysis?
the default is TRUE}
}
\value{
A matrix included SNP, CHR, position, MAF, canonical correlation, chisq, and pvalue.
}
\description{
the function is performed multivariate analysis of genotype–phenotype association based on CCA,
which calls \code{cca} in the **yacca** package to do canonical correlation analysis.
}
\examples{
data(geno)
data(pheno)
gpa<-magpa(geno,pheno,pca = TRUE)
head(gpa)

}
