% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtxtool.R
\name{Meshextract}
\alias{Meshextract}
\title{Extract sub mesh}
\usage{
Meshextract(ref.m, idx)
}
\arguments{
\item{ref.m}{a reference tri-mesh of 3D face}

\item{idx}{the index vector for extraction region}
}
\value{
a tri-mesh of facial region
}
\description{
The function extracts the vertices and edges from the reference tri-mesh based on the index vector
and reconstructs the tri-mesh for the vertices and edges which were extracted.
}
\examples{
data(refface)
ref.m <- refface$ref.m
seg.idx<- refface$seg.idx
nose.idx<- which(seg.idx==10)
nose.m <- Meshextract(ref.m,nose.idx)


}
