% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtxtool.R
\name{Meshvolume}
\alias{Meshvolume}
\title{Calculate the volume of a mesh}
\usage{
Meshvolume(m)
}
\arguments{
\item{m}{a tri-mesh object}
}
\value{
a vector of the volume of the mesh
}
\description{
the function calculate the volume of each tetrahedron
which was made by tri-mesh and origin of coordinates of the tri-mesh
}
\examples{
data(refface)
ref.m <- refface$ref.m
face.vol <- Meshvolume(ref.m)


}
