% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpa.R
\name{pcapheno}
\alias{pcapheno}
\title{Principal Component Analysis with Horn's Parallel Analysis}
\usage{
pcapheno(pheno, iterations = 100)
}
\arguments{
\item{pheno}{A matrix with n rows of sample and m columns of features}

\item{iterations}{sets the number of iterations with a user specified whole number representing
the number of random data sets to be produced in the analysis. The default is 100.}
}
\value{
a list of objects with the pca of phenotypes and the results of parallel analysis
}
\description{
the function is implemented to automatically extract the principal components from
high-dimension data, which calls \code{paran} to performs Horn's parallel analysis
for evaluating the components retained in a principle component analysis.
}
\examples{
data(pheno)
paral<- pcapheno(pheno)

}
