% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Phenotype Data}
\format{A matrix with n rows of sample and m columns of features}
\usage{
data(pheno)
}
\description{
The data is a matrix (2000 rows, 300 columns) expanded by 2000 samples
with 100 three-dimensional coordinate in each sample.
It should be used in the \code{magpa} function to test multivariate correlation.
}
\examples{
data(pheno)
str(pheno)
}
\keyword{datasets}
