% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readobj.R
\name{readobj}
\alias{readobj}
\title{Read a Wavefront OBJ 3D scene file into a tri-mesh}
\usage{
readobj(obj.file, mesh = TRUE)
}
\arguments{
\item{obj.file}{Path to an OBJ file}

\item{mesh}{boolean, construct tri-mesh? the default is TRUE}
}
\value{
a named list with items \code{shapes} and \code{materials},
  each containing sublists with one entry
  per object (shapes) or material (materials).
  Objects in the \code{shapes} list have the following structure \itemize{

  \item positions 3xN set of 3D vertices

  \item normals 3xN set of normal directions for each vertex (has 3 rows and
  0 cols when normals are not available)

  \item texcoords vector containing unprocessed texture coordinates

  \item indices 3/4xM set of indices into vertex array (trimesh/quadmesh)
  0-indexed

  \item material_ids (0-indexed, -1 when not set) }
}
\description{
This function reads a OBJ file, which a file format that is commonly used in 3D graphics
applications. It does not represent text, but does represent points, lines, polygons.
}
\examples{
data(refface)
ref.m <- refface$ref.m
obj.file <- tempfile(fileext = '.obj')
writeobj(obj.file, ref.m)
ref.m <- readobj(obj.file)

}
