% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refface.R
\docType{data}
\name{refface}
\alias{refface}
\title{Reference face}
\usage{
data(refface)
}
\description{
refface includes two object, a 3D reference face `ref.m` and
a label vecter of facial segmentation `seg.idx`.
}
\details{
\itemize{
  \item ref.m a large mesh3d object
  \item seg.idx a vector with 32251 length
}
}
\examples{
data(refface)
ref.m <- refface$ref.m
seg.idx<- refface$seg.idx
table(seg.idx)
visual3d(ref.m)
}
\keyword{datasets}
